Avoid endless loop while deleting

+++ src/Protocol/ActivityPub/Queue.php
@@ -128,6 +128,9 @@ class Queue
 
 		$children = DBA::select('inbox-entry', ['id'], ['in-reply-to-id' => $entry['object-id']]);
 		while ($child = DBA::fetch($children)) {
+			if ($id == $child['id']) {
+				continue;
+			}
 			self::deleteById($child['id']);
 		}
 		DBA::close($children);
